refactor: explode modules into separate bricks
